+++ priam/src/main/java/com/netflix/priam/notification/BackupNotificationMgr.java
@@ -73,11 +73,11 @@ public class BackupNotificationMgr implements EventObserver<BackupEvent> {
             jsonObject.put("region", instanceInfo.getRegion());
             jsonObject.put("rack", instanceInfo.getRac());
             jsonObject.put("token", instanceIdentity.getInstance().getToken());
-            jsonObject.put("backuptype", "SNAPSHOT_VERIFIED");
+            jsonObject.put(
+                    "backuptype", AbstractBackupPath.BackupFileType.SNAPSHOT_VERIFIED.name());
             jsonObject.put("snapshotInstant", backupVerificationResult.snapshotInstant);
             // SNS Attributes for filtering messages. Cluster name and backup file type.
-            Map<String, MessageAttributeValue> messageAttributes =
-                    getMessageAttributes(AbstractBackupPath.BackupFileType.SNAPSHOT_VERIFIED);
+            Map<String, MessageAttributeValue> messageAttributes = getMessageAttributes(jsonObject);
 
             this.notificationService.notify(jsonObject.toString(), messageAttributes);
         } catch (JSONException exception) {
@@ -88,21 +88,21 @@ public class BackupNotificationMgr implements EventObserver<BackupEvent> {
         }
     }
 
-    private Map<String, MessageAttributeValue> getMessageAttributes(
-            AbstractBackupPath.BackupFileType backupFileType) {
-        // SNS Attributes for filtering messages. Cluster name and backup file type.
-        Map<String, MessageAttributeValue> messageAttributes = new HashMap<>();
-        messageAttributes.putIfAbsent(
-                "s3clustername",
-                new MessageAttributeValue()
-                        .withDataType("String")
-                        .withStringValue(config.getAppName()));
-        messageAttributes.putIfAbsent(
-                "backuptype",
-                new MessageAttributeValue()
-                        .withDataType("String")
-                        .withStringValue(backupFileType.name()));
-        return messageAttributes;
+    private Map<String, MessageAttributeValue> getMessageAttributes(JSONObject message)
+            throws JSONException {
+        Map<String, MessageAttributeValue> attributes = new HashMap<>();
+        attributes.put("s3clustername", toStringAttribute(message.getString("s3clustername")));
+        attributes.put("backuptype", toStringAttribute(message.getString("backuptype")));
+        for (String attr : backupRestoreConfig.getBackupNotificationAdditionalMessageAttrs()) {
+            if (message.has(attr)) {
+                attributes.put(attr, toStringAttribute(String.valueOf(message.get(attr))));
+            }
+        }
+        return attributes;
+    }
+
+    private MessageAttributeValue toStringAttribute(String value) {
+        return new MessageAttributeValue().withDataType("String").withStringValue(value);
     }
 
     private void notify(AbstractBackupPath abp, String uploadStatus) {
@@ -127,10 +127,12 @@ public class BackupNotificationMgr implements EventObserver<BackupEvent> {
                 jsonObject.put("uploadstatus", uploadStatus);
                 jsonObject.put("compression", abp.getCompression().name());
                 jsonObject.put("encryption", abp.getEncryption().name());
+                jsonObject.put("isincremental", abp.isIncremental());
 
                 // SNS Attributes for filtering messages. Cluster name and backup file type.
                 Map<String, MessageAttributeValue> messageAttributes =
-                        getMessageAttributes(abp.getType());
+                        getMessageAttributes(jsonObject);
+
                 this.notificationService.notify(jsonObject.toString(), messageAttributes);
             } else {
                 logger.debug(