+++ priam/src/test/java/com/netflix/priam/config/FakeBackupRestoreConfig.java
@@ -13,6 +13,8 @@
  */
 package com.netflix.priam.config;
 
+import com.google.common.collect.ImmutableSet;
+
 /** Created by aagrawal on 6/26/18. */
 public class FakeBackupRestoreConfig implements IBackupRestoreConfig {
     @Override
@@ -28,5 +30,15 @@ public class FakeBackupRestoreConfig implements IBackupRestoreConfig {
     @Override
     public boolean enableV2Restore() {
         return false;
+    }
+
+    @Override
+    public int getBackupTTLMonitorPeriodInSec() {
+        return 0; // avoids sleeping altogether in tests.
+    }
+
+    @Override
+    public ImmutableSet<String> getBackupNotificationAdditionalMessageAttrs() {
+        return ImmutableSet.of();
     }
 }